Hoist Slot Leader dependencies up to BankingStage
